test module spinout
