--- litmus/clustering.py
@@ -1,36 +1,0 @@
-'''
-Placeholder file for running clustering algorithms
-
-todo - THIS FILE DEPRECATED
-
-Currently not in use as of 14/8/24
-
--HM
-'''
-
-# ========================
-import numpy as np
-from sklearn.cluster import DBSCAN
-from ._utils import *
-
-
-# ========================
-def clusterfind_1D(X: [float]) -> [int]:
-    '''
-    For some array of samples 'X', associate them with separated clusters
-    :param X:
-    :return:
-    '''
-    # todo - fill this out
-    return (np.ones_like(X))
-
-
-def sort_by_cluster(samples: dict, indices: [int]):
-    '''
-    A utility to break a dictionary up by some clustering method
-    :param samples:
-    :param indices:
-    :return:
-    '''
-    # todo - fill this out
-    return samples
