--- litmus/main.py
@@ -1,6 +1,0 @@
-from litmus import *
-
-import . lightcurve
-import . fitting_methods
-import . mocks
-import . models
